pyproject.toml
[tool.poetry]
name = "engine"
version = "0.1.0"
description = ""
authors = ["kaponda mulambia <44966170+kmulambia@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.6"}
uvicorn = "^0.32.1"
pydantic = "^2.10.3"
asyncpg = "^0.30.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
fastapi-limiter = "^0.1.6"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
setuptools = "^75.6.0"
pyjwt = "^2.10.1"
pytest-asyncio = "^0.24.0"
postmarker = "^1.0"
aio-pika = "^9.5.5"
pandas = "^2.2.3"
aiohttp = "^3.11.16"
africastalking = "^1.2.9"
openpyxl = "^3.1.5"
concurrent-log-handler = "^0.9.25"
pyodbc = "^5.2.0"
boto3 = "^1.38.1"
oracledb = "^3.1.0"
sshtunnel = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
